+++ xtp/src/libxtp/factories/toolfactory.cc
@@ -25,6 +25,7 @@
 #include "tools/coupling.h"
 #include "tools/densityanalysis.h"
 #include "tools/dftgwbse.h"
+#include "tools/diabatization.h"
 #include "tools/excitoncoupling.h"
 #include "tools/gencube.h"
 #include "tools/gpu_benchmark.h"
@@ -35,6 +36,7 @@
 #include "tools/orb2mol.h"
 #include "tools/partialcharges.h"
 #include "tools/spectrum.h"
+
 namespace votca {
 namespace xtp {
 
@@ -52,6 +54,7 @@ void QMToolFactory::RegisterAll(void) {
   QMTools().Register<APDFT>("apdft");
   QMTools().Register<Mol2Orb>("mol2orb");
   QMTools().Register<Orb2Mol>("orb2mol");
+  QMTools().Register<Diabatization>("diabatization");
   QMTools().Register<Orb2Fchk>("orb2fchk");
   QMTools().Register<GPUBenchmark>("gpu_benchmark");
 }