+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ Version 2023-dev
 -  unitary optimization for localization of orbitals (#971)
 -  separate option for density max convergence tolerance in qmmm (#987)
 -  Correct ECP charges added in Orca Molden reader (#990)
+-  Implemented ER diabatization for BSE (#988)
 -  update esp++ spce tutorial to py3 (#997)
 
 Version 2022.1 (released XX.01.22)
@@ -27,6 +28,7 @@ Version 2022.1 (released XX.01.22)
 -  fix CI builds with icpx (#984, #1004)
 -  fix CI builds with ESPResSo 4.2.0 (#993)
 -  fix localization issue (#998, #1000)
+-  update github actions to new versions (#1007)
 
 Version 2022 (released 15.01.22)
 ================================